Cache both riscv-tools and verilator for faster CI tests

+++ ci/prepare-riscv-tools.sh
@@ -6,11 +6,13 @@
 set -x
 set -e
 
-# move to top level dir and remove boom-template if present
-cd ..
-rm -rf boom-template
+if [ ! -d "/home/circleci/riscv-tools-install" ]; then
+    # move to top level dir and remove boom-template if present
+    cd ..
+    rm -rf boom-template
 
-# clone boom-template and create the riscv-tools
-git clone --progress --verbose https://github.com/riscv-boom/boom-template.git
-cd boom-template
-./scripts/build-tools.sh
+    # clone boom-template and create the riscv-tools
+    git clone --progress --verbose https://github.com/riscv-boom/boom-template.git
+    cd boom-template
+    ./scripts/build-tools.sh
+fi
